+++ test/constructorParams.js
@@ -1,14 +1,14 @@
 /* eslint-env mocha */
 
 const assert = require('assert')
-const cleanupTestApp = require('../util/cleanupTestApp')
+const cleanupTestApp = require('./util/cleanupTestApp')
 const fs = require('fs-extra')
 const path = require('path')
 
 describe('Constructor Parameter Tests', function () {
-  const appDir = path.join(__dirname, '../app/constructorParams')
-  const config = require('../util/testConstructorConfig.json')
-  const pkgConfig = require('../util/testPkgConfig.json')
+  const appDir = path.join(__dirname, 'app/constructorParams')
+  const config = require('./util/testConstructorConfig.json')
+  const pkgConfig = require('./util/testPkgConfig.json')
   const params = Object.keys(config)
   const pkg = {
     rooseveltConfig: pkgConfig
@@ -19,7 +19,7 @@ describe('Constructor Parameter Tests', function () {
     fs.ensureDirSync(path.join(appDir))
     fs.writeFileSync(path.join(appDir, 'package.json'), JSON.stringify(pkg))
 
-    app = require('../../roosevelt')({
+    app = require('../roosevelt')({
       appDir: appDir,
       enableCLIFlags: false,
       ...config